+++ src/aiida_vasp/commands/__init__.py
@@ -28,5 +28,7 @@ def cmd_aiida_vasp() -> None:
 # These lines simply triggers python to include the sub-commands
 from .immigrant import import_calc
 from .launch import launch_workchain
+from .potcar import potcar
+from .tools import export
 
-__all__ = ['cmd_aiida_vasp', 'import_calc', 'launch_workchain']
+__all__ = ['cmd_aiida_vasp', 'export', 'import_calc', 'launch_workchain', 'potcar']